[ADD] one more test for long numbers

+++ types.hpp
@@ -15,7 +15,7 @@
 #ifndef TYPES_HPP
 #define TYPES_HPP
 
-typedef long int int64;
+typedef long long int int64;
 
 
 #endif
